+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Moved IC342 scheduling range to configuration.
 - Add tool to set IC342 LST range.
+- By default, prioritise incomplete plates already observed that night (planner).
 
 
 ## [2.2.0] - 2019-09-30